add .travis.yml configuration
We build the wheels and run the tests on both Linux and OSX, each on Python 2.7, 3.5 and 3.6, 32 and 64 bit, narrow and wide Unicode builds... phew!

Most of this is boilerplate from https://github.com/matthew-brett/multibuild/blob/master/README.rst

The auto-deployment to Github and PyPI is currently disabled, as it requires Adam to fill in the encrypted tokens.