[setup.py] remove -ansi -pedantic /Za compiler flags
Enforcing ANSI C creates problems when building extension for Python 3.6,
because of the presence of C++ style one-line comments in some of the python header files.

Build failure:
https://travis-ci.org/anthrotype/obp-zopfli/jobs/193466619#L370-L371

Here's a relevant issue on the Python bug tracker:
https://bugs.python.org/issue29215

Apparently, Python 3.6 now requires some C99 support in the toolchain:
https://docs.python.org/3/whatsnew/3.6.html#build-and-c-api-changes

If we disable -ansi -pedantic, the 3.6 build passes and the compiled extension seems to work just fine.

For consistency, I also dropped /Za which does the same thing for visual studio, and the build/test also pass without issues without it.

I only kept -lm for "unix" compiler (to explicitly link the C math library).
I haven't tested cygwin or mingw32 on Windows, so I don't know whether that -lm linker flag is required for them or not. So in doubt, I don't pass it.