[setup.py] add 'pass' command that does nothing...
For Travis CI PyPI 'deploy' stage, we must provide 'distributions' to build (e.g. sdist, bdist_wheel) and upload to PyPI.

However, by the time Travis deploys, we have already built the wheels (for Linux it's important we do it in the manylinux1 Docker container), and we want to upload those, instead of rebuilding them.
Similarly, the sdist we want to build/upload it only once, and we do that in the 'after_success' stage if BUILD_SDIST is true.

I know it feels like a hack, but that's the only workaround I could find.
Alternatively, we could scrap Travis auto-deploy feature and call the 'twine' command ourselves (which is what we do for Appveyor which does not provide the PyPI deploy feauture).