[setup.py] add .post1 suffix to version string; can't re-upload same files to PyPI
The previous deployment failed so I had to delete 0.1.0 tag.
But in the meantime some files had already been uploaded to PyPI for 0.1.0.
PyPI does not let you upload a file with the same name twice, so we need to
distinguish this post-release from the 0.1.0 one.

I don't want to bump the patch version just for this (e.g. 0.1.1).
I post release is better, it's meant for this kind of things.